[update] middleware system integration testing.
modulize `Connection` and `Visualization`. The `system.py` which
holds all the processes and threads will be the core of the system.
Three Queues are used to get data or command message to different
process or thread. The upgraded `client_test.py` will send simulat-
ed yaw and pitch data for testing.